move ConstantsClass and make it more widely used
